fuse_create.cpp

+++ src/fuse_create.cpp
@@ -287,7 +287,7 @@ namespace FUSE
     if(rv != 0)
       return rv;
 
-    uint64_t x = ((+cfg->passthrough << 32));
+    uint64_t x = ((((uint64_t)+cfg->passthrough) << 32));
     switch(+cfg->passthrough)
       {
       case PassthroughEnum::off:
